ADD init

+++ protocol/contracts/libraries/SvgUtils.sol
@@ -1,63 +1,79 @@
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.4;
-import "@openzeppelin/contracts/utils/Strings.sol";
+import '@openzeppelin/contracts/utils/Strings.sol';
 
 library SvgUtils {
   using Strings for uint256;
 
-  uint public constant DECIMALS = 4;
-  uint public constant ONE_UNIT = 10 ** DECIMALS;
+  uint256 public constant DECIMALS = 4;
+  uint256 public constant ONE_UNIT = 10**DECIMALS;
 
-  function padZeros(string memory s, uint len) public pure returns (string memory) {
-    uint local_len = bytes(s).length;
+  function padZeros(string memory s, uint256 len)
+    public
+    pure
+    returns (string memory)
+  {
+    uint256 local_len = bytes(s).length;
     string memory local_s = s;
-    while(local_len < len) {
+    while (local_len < len) {
       local_s = string(abi.encodePacked('0', local_s));
       local_len++;
     }
     return local_s;
   }
 
-  function wholeNumber(uint n) public pure returns (uint) {
+  function wholeNumber(uint256 n) public pure returns (uint256) {
     return n / ONE_UNIT;
   }
 
-  function decimals(uint n) public pure returns (uint) {
+  function decimals(uint256 n) public pure returns (uint256) {
     return n % ONE_UNIT;
   }
 
-  function toDecimalString(uint n) public pure returns (string memory s) {
+  function toDecimalString(uint256 n) public pure returns (string memory s) {
     if (n == 0) return '0';
 
-    s = string(abi.encodePacked(
-      (n / (ONE_UNIT)).toString(), '.' , padZeros((n % ONE_UNIT).toString(), DECIMALS)
-    ));
+    s = string(
+      abi.encodePacked(
+        (n / (ONE_UNIT)).toString(),
+        '.',
+        padZeros((n % ONE_UNIT).toString(), DECIMALS)
+      )
+    );
   }
 
-  function lerpWithDecimals(uint min, uint max, bytes1 scale) public pure returns (uint) {
+  function lerpWithDecimals(
+    uint256 min,
+    uint256 max,
+    bytes1 scale
+  ) public pure returns (uint256) {
     if (scale == 0x0) return min * ONE_UNIT;
     if (scale == 0xff) return max * ONE_UNIT;
-    uint delta = ((max - min) * ONE_UNIT * uint(uint8(scale))) / 255; 
+    uint256 delta = ((max - min) * ONE_UNIT * uint256(uint8(scale))) / 255;
     return (min * ONE_UNIT) + delta;
   }
 
-  bytes16 private constant _HEX_SYMBOLS = "0123456789abcdef";
+  bytes16 private constant _HEX_SYMBOLS = '0123456789abcdef';
 
   function toColorHexString(uint256 value) public pure returns (string memory) {
-      bytes memory buffer = new bytes(2 * 3 + 1);
-      buffer[0] = "#";
-      for (uint256 i = 2 * 3; i > 0; --i) {
-          buffer[i] = _HEX_SYMBOLS[value & 0xf];
-          value >>= 4;
-      }
-      require(value == 0, "Strings: hex length insufficient");
-      return string(buffer);
+    bytes memory buffer = new bytes(2 * 3 + 1);
+    buffer[0] = '#';
+    for (uint256 i = 2 * 3; i > 0; --i) {
+      buffer[i] = _HEX_SYMBOLS[value & 0xf];
+      value >>= 4;
+    }
+    require(value == 0, 'Strings: hex length insufficient');
+    return string(buffer);
   }
 
-  function toColorHexStringByBytes(bytes1 r, bytes1 g, bytes1 b) public pure returns (string memory) {
+  function toColorHexStringByBytes(
+    bytes1 r,
+    bytes1 g,
+    bytes1 b
+  ) public pure returns (string memory) {
     bytes memory buffer = new bytes(7);
-    buffer[0] = "#";
+    buffer[0] = '#';
     buffer[2] = _HEX_SYMBOLS[uint8(r) & 0xf];
     r >>= 4;
     buffer[1] = _HEX_SYMBOLS[uint8(r) & 0xf];
@@ -69,9 +85,12 @@ library SvgUtils {
     buffer[5] = _HEX_SYMBOLS[uint8(b) & 0xf];
     return string(buffer);
   }
-  
-  function toColorHexStringByBytes3(bytes3 rgb) public pure returns (string memory) {
+
+  function toColorHexStringByBytes3(bytes3 rgb)
+    public
+    pure
+    returns (string memory)
+  {
     return toColorHexStringByBytes(rgb[0], rgb[1], rgb[2]);
   }
-
 }
